break up useMultichainBalances hook
