+++ app/components/Views/confirmations/components/rows/account-network-info-row/account-network-info-expanded/account-network-info-expanded.tsx
@@ -9,7 +9,7 @@ import Network from '../../../UI/info-row/info-value/network';
 import { useSignatureRequest } from '../../../../hooks/signatures/useSignatureRequest';
 import { Hex } from '@metamask/utils';
 import { renderShortAddress } from '../../../../../../../util/address';
-import { useMultichainBalances } from '../../../../../../hooks/useMultichainBalances';
+import { useMultichainBalancesForSelectedAccount } from '../../../../../../hooks/useMultichainBalances';
 import { useTransactionMetadataRequest } from '../../../../hooks/transactions/useTransactionMetadataRequest';
 
 const AccountNetworkInfoExpanded = () => {
@@ -26,7 +26,8 @@ const AccountNetworkInfoExpanded = () => {
     fromAddress = transactionMetadata?.txParams?.from as string;
   }
   const { accountAddress } = useAccountInfo(fromAddress);
-  const { selectedAccountMultichainBalance } = useMultichainBalances();
+  const { selectedAccountMultichainBalance } =
+    useMultichainBalancesForSelectedAccount();
   const balanceToDisplay = selectedAccountMultichainBalance?.displayBalance;
 
   return (
