+++ app/components/hooks/useMultichainBalances/index.ts
@@ -1,5 +1,6 @@
-export { default as useMultichainBalances } from './useMultichainBalances';
+export { default as useMultichainBalancesForAllAccounts } from './useMultichainBalancesForAllAccounts';
+export { default as useMultichainBalancesForSelectedAccount } from './useMultichainBalancesForSelectedAccount';
 export type {
-  UseMultichainBalancesHook,
+  UseSelectedAccountMultichainBalancesHook,
   MultichainBalancesData,
 } from './useMultichainBalances.types';
