+++ app/components/hooks/useMultichainBalances/useMultichainBalancesForAllAccounts.ts
@@ -1,0 +1,138 @@
+/* eslint-disable arrow-body-style */
+import { useSelector } from 'react-redux';
+import { isPortfolioViewEnabled } from '../../../util/networks';
+import {
+  selectChainId,
+  selectIsPopularNetwork,
+  selectProviderConfig,
+  selectEvmTicker,
+  selectEvmChainId,
+} from '../../../selectors/networkController';
+import { selectCurrentCurrency } from '../../../selectors/currencyRateController';
+import { selectIsTokenNetworkFilterEqualCurrentNetwork } from '../../../selectors/preferencesController';
+import { selectInternalAccounts } from '../../../selectors/accountsController';
+import { getChainIdsToPoll } from '../../../selectors/tokensController';
+import { useGetFormattedTokensPerChain } from '../useGetFormattedTokensPerChain';
+import { useGetTotalFiatBalanceCrossChains } from '../useGetTotalFiatBalanceCrossChains';
+import useIsOriginalNativeTokenSymbol from '../useIsOriginalNativeTokenSymbol/useIsOriginalNativeTokenSymbol';
+import {
+  MultichainBalancesData,
+  UseAllAccountsMultichainBalancesHook,
+} from './useMultichainBalances.types';
+///: BEGIN:ONLY_INCLUDE_IF(keyring-snaps)
+import {
+  selectMultichainShouldShowFiat,
+  selectMultichainBalances,
+  selectMultichainAssets,
+  selectMultichainAssetsRates,
+} from '../../../selectors/multichain';
+import { selectSelectedNonEvmNetworkChainId } from '../../../selectors/multichainNetworkController';
+///: END:ONLY_INCLUDE_IF
+import { useMemo } from 'react';
+import {
+  getAccountBalanceData,
+  getAggregatedBalance,
+  getShouldShowAggregatedPercentage,
+} from './utils';
+import { SupportedCaipChainId } from '@metamask/multichain-network-controller';
+/**
+ * Hook to manage portfolio balance data across chains.
+ *
+ * @returns Portfolio balance data
+ */
+const useMultichainBalancesForAllAccounts =
+  (): UseAllAccountsMultichainBalancesHook => {
+    const accountsList = useSelector(selectInternalAccounts);
+    const chainId = useSelector(selectChainId);
+    const evmChainId = useSelector(selectEvmChainId);
+    const currentCurrency = useSelector(selectCurrentCurrency);
+    const allChainIDs = useSelector(getChainIdsToPoll);
+    const isTokenNetworkFilterEqualCurrentNetwork = useSelector(
+      selectIsTokenNetworkFilterEqualCurrentNetwork,
+    );
+    const isPopularNetwork = useSelector(selectIsPopularNetwork);
+    const { type } = useSelector(selectProviderConfig);
+    const ticker = useSelector(selectEvmTicker);
+
+    // Production hooks (EVM)
+    const formattedTokensWithBalancesPerChain = useGetFormattedTokensPerChain(
+      accountsList,
+      !isTokenNetworkFilterEqualCurrentNetwork && isPopularNetwork,
+      allChainIDs,
+    );
+
+    const totalFiatBalancesCrossEvmChain = useGetTotalFiatBalanceCrossChains(
+      accountsList,
+      formattedTokensWithBalancesPerChain,
+    );
+
+    const isOriginalNativeEvmTokenSymbol = useIsOriginalNativeTokenSymbol(
+      evmChainId,
+      ticker,
+      type,
+    );
+
+    ///: BEGIN:ONLY_INCLUDE_IF(keyring-snaps)
+    const shouldShowFiat = useSelector(selectMultichainShouldShowFiat);
+    const multichainBalances = useSelector(selectMultichainBalances);
+    const multichainAssets = useSelector(selectMultichainAssets);
+    const multichainAssetsRates = useSelector(selectMultichainAssetsRates);
+    const nonEvmChainId = useSelector(selectSelectedNonEvmNetworkChainId);
+    ///: END:ONLY_INCLUDE_IF
+
+    const isPortfolioEnabled = isPortfolioViewEnabled();
+
+    // Create a stable reference for each account's balance data
+    const allAccountBalances = useMemo(() => {
+      const result: Record<string, MultichainBalancesData> = {};
+
+      for (const account of accountsList) {
+        const accountBalanceData = getAccountBalanceData(
+          account,
+          multichainBalances,
+          multichainAssets,
+          multichainAssetsRates,
+          nonEvmChainId,
+          shouldShowFiat,
+          currentCurrency,
+          totalFiatBalancesCrossEvmChain,
+          isOriginalNativeEvmTokenSymbol,
+        );
+        result[account.id] = {
+          displayBalance: accountBalanceData.displayBalance,
+          displayCurrency: currentCurrency,
+          totalFiatBalance: accountBalanceData.totalFiatBalance,
+          totalNativeTokenBalance: accountBalanceData.totalNativeTokenBalance,
+          nativeTokenUnit: accountBalanceData.nativeTokenUnit,
+          tokenFiatBalancesCrossChains:
+            totalFiatBalancesCrossEvmChain[account.address]
+              ?.tokenFiatBalancesCrossChains ?? [],
+          shouldShowAggregatedPercentage: getShouldShowAggregatedPercentage(
+            account,
+            chainId as SupportedCaipChainId,
+          ),
+          isPortfolioVieEnabled: isPortfolioEnabled,
+          aggregatedBalance: getAggregatedBalance(account),
+        };
+      }
+
+      return result;
+    }, [
+      accountsList,
+      chainId,
+      currentCurrency,
+      isOriginalNativeEvmTokenSymbol,
+      isPortfolioEnabled,
+      multichainAssets,
+      multichainAssetsRates,
+      multichainBalances,
+      nonEvmChainId,
+      shouldShowFiat,
+      totalFiatBalancesCrossEvmChain,
+    ]);
+    return {
+      multichainBalancesForAllAccounts: allAccountBalances,
+    };
+  };
+
+export default useMultichainBalancesForAllAccounts;
